Fix classpath issues
There was an issue with the JAR's classpath not matching the names of the files in the `lib` directory because the classpath's files had their versions after their names whereas the actual files did not, so I changed the files to include their versions.
Also, Nimrod L&F wasn't included in the classpath at all.