Fix translation errors and missing documentation
I found the original Shimeji source and decided to add all of its documentation by translating it. Because Google Translate is much better now than it was when this was originally forked, these doc comments are much easier to understand.
I also translated the original README, and discovered that it said to keep the license in Japanese, so I did that.
I also edited the credits so they are more accurate to who actually worked on which file. For instance, Yuki was credited for several files which were actually created entirely by the Shimeji-ee Group.