Move schema language files to resource directory
These files shouldn't be customized by the user, so it's better to bundle them with the JAR to prevent any modifications. This also lets us preload them instead of loading them every time any XML file is being read.
I may consider moving more files (specifically the language and icon files) to the resource directory, but I want to implement ways for the user to override them to preserve the existing customizability.